PR-FIX: make the CloudEvent parsing part of the abstraction
