+++ src/Arcus.WebApi.Jobs/KeyVault/AzureServiceBusTopicCloudEventSubscriptionMessagePump.cs
@@ -1,23 +1,29 @@
 ﻿using System;
+using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 using Arcus.Messaging.Abstractions;
 using Arcus.Messaging.Pumps.ServiceBus;
+using CloudNative.CloudEvents;
+using GuardNet;
 using Microsoft.Azure.ServiceBus;
 using Microsoft.Azure.ServiceBus.Management;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
+using Newtonsoft.Json.Linq;
 
 namespace Arcus.WebApi.Jobs.KeyVault
 {
     /// <summary>
     /// Representing a Azure ServiceBus Topic message pump that will create and delete a ServiceBus Topic subscription during the lifetime of the pump.
     /// </summary>
-    public abstract class TempSubscriptionAzureServiceBusMessagePump<TMessage> : AzureServiceBusMessagePump<TMessage>
+    public abstract class AzureServiceBusTopicCloudEventSubscriptionMessagePump : AzureServiceBusMessagePump<CloudEvent>
     {
         private readonly string _topicPath, _subscriptionName;
         private readonly ManagementClient _managementClient;
+
+        private static readonly JsonEventFormatter JsonEventFormatter = new JsonEventFormatter();
 
         /// <summary>
         /// Constructor
@@ -25,7 +31,7 @@ namespace Arcus.WebApi.Jobs.KeyVault
         /// <param name="configuration">Configuration of the application</param>
         /// <param name="serviceProvider">Collection of services that are configured</param>
         /// <param name="logger">Logger to write telemetry to</param>
-        protected TempSubscriptionAzureServiceBusMessagePump(
+        protected AzureServiceBusTopicCloudEventSubscriptionMessagePump(
             IConfiguration configuration,
             IServiceProvider serviceProvider,
             ILogger logger) : base(configuration, serviceProvider, logger)
@@ -40,9 +46,24 @@ namespace Arcus.WebApi.Jobs.KeyVault
             _managementClient = new ManagementClient(connectionString);
         }
 
+        /// <summary>
+        /// Deserializes a raw JSON message body.
+        /// </summary>
+        /// <param name="rawMessageBody">Raw message body to deserialize</param>
+        /// <param name="messageContext">Context concerning the message</param>
+        /// <returns>Deserialized message</returns>
+        protected override CloudEvent DeserializeJsonMessageBody(byte[] rawMessageBody, MessageContext messageContext)
+        {
+            Guard.NotNull(rawMessageBody, nameof(rawMessageBody), "Cannot deserialize raw JSON body from 'null' input");
+            Guard.NotAny(rawMessageBody, nameof(rawMessageBody), "Cannot deserialize raw JSON body from empty input");
+
+            CloudEvent cloudEvent = JsonEventFormatter.DecodeStructuredEvent(rawMessageBody);
+            return cloudEvent;
+        }
+
         /// <inheritdoc />
         protected abstract override Task ProcessMessageAsync(
-            TMessage message,
+            CloudEvent message,
             AzureServiceBusMessageContext messageContext,
             MessageCorrelationInfo correlationInfo,
             CancellationToken cancellationToken);
